fix: add retry mechanism when downloading

+++ triedb/pathdb/async_incr_state.go
@@ -215,7 +215,7 @@ func (c *incrNodeCache) flushToAncientDB(incrDB *rawdb.IncrSnapDB) error {
 		if len(entry.Nodes) > 0 {
 			jn = append(jn, entry)
 			entrySize := ownerSize + nodesListSize
-			totalSize += uint64(entrySize)
+			totalSize += entrySize
 		}
 
 		if totalSize >= c.batchSize {
